requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1<0.16
pillow==10.0.0
django-storages[google]
google-auth
google-api-python-client
drf_yasg==1.21.7
